+++ lib/Mood/Mood.cpp
@@ -9,6 +9,7 @@
 #include "KeyBoard.h"
 
 #include "Affichage.h"
+#include "AppConfig.h"
 
 #include <WiFiClient.h>
 #include <WebServer.h>
@@ -27,10 +28,6 @@ String version = "0.0.4" ;
 unsigned long loopCount;
 unsigned long startTime;
 String msg;
-
-// Replace with your network credentials
-const char *ssid = "Mood1";
-const char *password = "123456789";
 
 // Set web server port number to 80
 WiFiServer server(80);
@@ -117,7 +114,9 @@ void process_keyPressed(char key_pressed) {
 
   if (isBoutonVote(key_pressed)) {
     if (selected_candidat == -1) {
-      selected_candidat = 0;
+      Serial.println("Veuillez selectionner un candidat d'abord");
+      affichage.afficherErreurDeSaisie();
+      return;
     } 
     gererVote(getMood(key_pressed));
   } else {
@@ -154,8 +153,8 @@ void onOTAEnd(bool success) {
 
 void Mood::init()
 {
-  affichage.initAffichage(false);
   Serial.begin(115200);
+  affichage.initAffichage(false);
   initRTC();
   initAp();
 
@@ -164,6 +163,11 @@ void Mood::init()
 
   votes_en_cours.init(candidats, MAX_CANIDATS);
   clavier.initKeyboard(true);
+  if (!preferences.begin("mood-stats", false)) {
+    Serial.println("Impossible d'ouvrir les preferences");
+  } else {
+    restore_vote_en_cours();
+  }
   
   server.begin();
 
@@ -171,10 +175,6 @@ void Mood::init()
 
   Serial.print("Setup Complete !");
   affichage.afficherInitEnCours();
-
-  preferences.begin("mood-stats", false); 
-
-
 }
 
 
@@ -183,10 +183,18 @@ void initOTA(){
   Serial.println("Start OTA initialization...");
 
   ota_server.on("/", []() {
+    if (!ota_server.authenticate(AppConfig::webUser, AppConfig::webPassword)) {
+      ota_server.requestAuthentication();
+      return;
+    }
     ota_server.send(200, "text/plain", "Hi! This is ElegantOTA Demo Oh yeaaahh !!!. version "+version);
   });
 
   ota_server.on("/download", HTTP_GET, []() {
+    if (!ota_server.authenticate(AppConfig::webUser, AppConfig::webPassword)) {
+      ota_server.requestAuthentication();
+      return;
+    }
     String s = "Version;" + String(version);
     s+= "----------;------------;--------------;------------\n";
     s += "Candidat;Heureux;Indifferent;Triste\n";
@@ -219,22 +227,35 @@ void initOTA(){
   });
 
   ota_server.on("/save", HTTP_GET, []() {
+    if (!ota_server.authenticate(AppConfig::webUser, AppConfig::webPassword)) {
+      ota_server.requestAuthentication();
+      return;
+    }
     save_vote_en_cours();
     ota_server.send(200, "text/plain", "Hi, vote en cours sauvegardé. ");    
   });
 
   ota_server.on("/restore", HTTP_GET, []() {
+    if (!ota_server.authenticate(AppConfig::webUser, AppConfig::webPassword)) {
+      ota_server.requestAuthentication();
+      return;
+    }
     restore_vote_en_cours();
     ota_server.send(200, "text/plain", "Hi, vote en cours restaurés. ");    
   });
 
   ota_server.on("/reset", HTTP_GET, []() {
+    if (!ota_server.authenticate(AppConfig::webUser, AppConfig::webPassword)) {
+      ota_server.requestAuthentication();
+      return;
+    }
     preferences.clear();
+    votes_en_cours.resetVotes();
       Serial.println("Suppression de la sauvegarde en flash");
     ota_server.send(200, "text/plain", "memory cleared. ");    
   });
 
-  ElegantOTA.begin(&ota_server);    // Start ElegantOTA
+  ElegantOTA.begin(&ota_server, AppConfig::webUser, AppConfig::webPassword);    // Start ElegantOTA
   // ElegantOTA callbacks
   ElegantOTA.onStart(onOTAStart);
   ElegantOTA.onProgress(onOTAProgress);
@@ -249,7 +270,7 @@ void initAp(){
   Serial.print("Setting AP (Access Point)…");
   // Remove the password parameter, if you want the AP (Access Point) to be open
   //WiFi.softAP(ssid, password);
-  WiFi.softAP(ssid);
+  WiFi.softAP(AppConfig::wifiSsid, AppConfig::wifiPassword);
 
   IPAddress IP = WiFi.softAPIP();
   Serial.print("AP IP address: ");
@@ -287,6 +308,14 @@ void Mood::loop()
 }
 
 void gererVote(int mood) {
+  if (selected_candidat < 0 || selected_candidat >= MAX_CANIDATS ||
+      mood < 1 || mood > MAX_MOODS) {
+    Serial.println("Vote invalide");
+    affichage.afficherErreurDeSaisie();
+    selected_candidat = -1;
+    return;
+  }
+
   Serial.print("Vote pour ");
   Serial.print(candidats[selected_candidat]);
   Serial.print(" : ");
@@ -327,15 +356,38 @@ void initResponseOkCsv(WiFiClient & client){
 }
 
 String extractTimeStringFromHeader(String & header){
-  int start = header.indexOf("GET /set-time/?datetimestr=") + 27;
+  const int markerStart = header.indexOf("GET /set-time/?datetimestr=");
+  if (markerStart < 0) {
+    return String();
+  }
+  int start = markerStart + 27;
   int end = start + 14;
   return header.substring(start, end);
 }
 
-void handleSetTimeUrl(WiFiClient & client, String & header){
-  initResponseOk(client);
+bool isValidTimeString(const String & time){
+  if (time.length() != 14) {
+    return false;
+  }
+  for (unsigned int i = 0; i < time.length(); i++) {
+    if (!isDigit(time.charAt(i))) {
+      return false;
+    }
+  }
+  return true;
+}
 
+void handleSetTimeUrl(WiFiClient & client, String & header){
   String time = extractTimeStringFromHeader(header);  
+  if (!isValidTimeString(time)) {
+    client.println("HTTP/1.0 400 Bad Request");
+    client.println("Content-type:text/plain");
+    client.println("Connection: close");
+    client.println();
+    client.println("Invalid date format");
+    return;
+  }
+
   int year = time.substring(0, 4).toInt();
   int month = time.substring(4, 6).toInt();
   int day = time.substring(6, 8).toInt();
@@ -354,7 +406,19 @@ void handleSetTimeUrl(WiFiClient & client, String & header){
   Serial.println(min);
   Serial.print("sec : " );
   Serial.println(sec);
-  rtc.adjust(DateTime(year, month, day, hour, min, sec));
+  DateTime newDate(year, month, day, hour, min, sec);
+  if (!newDate.isValid()) {
+    client.println("HTTP/1.0 400 Bad Request");
+    client.println("Content-type:text/plain");
+    client.println("Connection: close");
+    client.println();
+    client.println("Invalid date value");
+    return;
+  }
+
+  rtc.adjust(newDate);
+  initResponseOk(client);
+  client.println("Date updated");
   
 }
 
@@ -398,14 +462,24 @@ void handleWifiClient(){
   if (client)
   {                                // If a new client connects,
     Serial.println("New Client."); // print a message out in the serial port
+    header = "";
+    header.reserve(1024);
     String currentLine = "";       // make a String to hold incoming data from the client
+    unsigned long lastActivity = millis();
     while (client.connected())
     { // loop while the client's connected
+      if (millis() - lastActivity > 2000) {
+        break;
+      }
       if (client.available())
       {                         // if there's bytes to read from the client,
         char c = client.read(); // read a byte, then
+        lastActivity = millis();
         Serial.write(c);        // print it out the serial monitor
         header += c;
+        if (header.length() > 1024) {
+          break;
+        }
         if (c == '\n')
         { // if the byte is a newline character
           // if the current line is blank, you got two newline characters in a row.
